Restructuring and rendering stuff
Worked on many things since the last push. Did a bit of restructuring to try and separate engine and game, for example. I'm thinking the goal is to sooner or later make the engine work as a framework where it will be an easy process to use the same engine for different types of game projects.

I brought in some code from a previous project/attempt and adapted some of it to fit this project.

Fixed an issue where the validation layer messages wouldn't log to the console. 

Wrote a parser that cleans up some Vulkan validation messages and makes them easier to read (Vulkan object errors in particular). 

Did some restructuring of the rendering-related Vulkan objects to have more of a central "context" object. By doing it this way, you can more easily access every handle related to Vulkan, rather than having a lot of it abstracted away under more subobjects. For example, the vk::SurfaceKHR and ash::Device were under yet another device object before. Now they're both simply under the context object. I also find it easier to follow the destruction order of Vulkan objects by doing it this way.

Decoupling the game and the engine ties into the renderer. I want to, later on, have a renderer frontend/backend type of structure where the boilerplate is abstracted away within the engine. The engine's users should be able to do something as simple as creating a mesh struct by providing a path to a model file, and it should be simple to associate that mesh with different shaders.

For now, I've decided to skip the separation of graphics queue and present queue and simplify it by only using the graphics queue for everything (I think they're the same on a lot of graphics cards anyway). I want to get something working before going too deep into the details of that. I haven't worked a lot with synchronization stuff before either, so I'm thinking having a single queue, to begin with, is probably going to be for the better.